// GarmentsManagementSystemAngular/src/app/components/features/production/cutting-plan/models/order-response.ts
export interface OrderResponse {


      id: number;

  orderId: string;

  poNumber: string;

  buyerId: number;

  buyerCode: string;

  buyerName: string;

  styleId: number;

  styleCode: string;

  styleName: string;

  orderDate: string;

  shipDate: string;

  status: string;

  shippingAddress: string;

  totalQuantity: number;

}